Move jex.ffi to jax.ffi.
