--- docs/jax.extend.ffi.rst
@@ -1,12 +1,0 @@
-``jax.extend.ffi`` module
-=========================
-
-.. automodule:: jax.extend.ffi
-
-.. autosummary::
-  :toctree: _autosummary
-
-  ffi_call
-  ffi_lowering
-  pycapsule
-  register_ffi_target
