+++ docs/jax.extend.rst
@@ -12,7 +12,6 @@ Modules
   :maxdepth: 1
 
   jax.extend.core
-  jax.extend.ffi
   jax.extend.linear_util
   jax.extend.mlir
   jax.extend.random
