+++ examples/ffi/src/jax_ffi_example/cpu_examples.py
@@ -15,28 +15,27 @@
 import numpy as np
 
 import jax
-import jax.extend as jex
 
 from jax_ffi_example import _cpu_examples
 
 for name, target in _cpu_examples.registrations().items():
-  jex.ffi.register_ffi_target(name, target)
+  jax.ffi.register_ffi_target(name, target)
 
 
 def array_attr(num: int):
-  return jex.ffi.ffi_call(
+  return jax.ffi.ffi_call(
       "array_attr",
       jax.ShapeDtypeStruct((), np.int32),
   )(array=np.arange(num, dtype=np.int32))
 
 
 def dictionary_attr(**kwargs):
-  return jex.ffi.ffi_call(
+  return jax.ffi.ffi_call(
       "dictionary_attr",
       (jax.ShapeDtypeStruct((), np.int32), jax.ShapeDtypeStruct((), np.int32)),
   )(**kwargs)
 
 
 def counter(index):
-  return jex.ffi.ffi_call(
+  return jax.ffi.ffi_call(
     "counter", jax.ShapeDtypeStruct((), jax.numpy.int32))(index=int(index))
