+++ examples/ffi/src/jax_ffi_example/cuda_examples.py
@@ -24,15 +24,14 @@ import numpy as np
 
 import jax
 import jax.numpy as jnp
-import jax.extend as jex
 
 # Load the shared library with the FFI target definitions
 SHARED_LIBRARY = os.path.join(os.path.dirname(__file__), "lib_cuda_examples.so")
 library = ctypes.cdll.LoadLibrary(SHARED_LIBRARY)
 
-jex.ffi.register_ffi_target("foo-fwd", jex.ffi.pycapsule(library.FooFwd),
+jax.ffi.register_ffi_target("foo-fwd", jax.ffi.pycapsule(library.FooFwd),
                             platform="CUDA")
-jex.ffi.register_ffi_target("foo-bwd", jex.ffi.pycapsule(library.FooBwd),
+jax.ffi.register_ffi_target("foo-bwd", jax.ffi.pycapsule(library.FooBwd),
                             platform="CUDA")
 
 
@@ -42,7 +41,7 @@ def foo_fwd(a, b):
   assert a.dtype == b.dtype
   n = np.prod(a.shape).astype(np.uint64)
   out_type = jax.ShapeDtypeStruct(a.shape, a.dtype)
-  c, b_plus_1 = jex.ffi.ffi_call("foo-fwd", (out_type, out_type))(a, b, n=n)
+  c, b_plus_1 = jax.ffi.ffi_call("foo-fwd", (out_type, out_type))(a, b, n=n)
   return c, (a, b_plus_1)
 
 
@@ -55,7 +54,7 @@ def foo_bwd(res, c_grad):
   assert a.dtype == b_plus_1.dtype
   n = np.prod(a.shape).astype(np.uint64)
   out_type = jax.ShapeDtypeStruct(a.shape, a.dtype)
-  return jex.ffi.ffi_call("foo-bwd", (out_type, out_type))(c_grad, a, b_plus_1,
+  return jax.ffi.ffi_call("foo-bwd", (out_type, out_type))(c_grad, a, b_plus_1,
                           n=n)
